// src/api/call-api-history/content-types/call-api-history/call-api-history.ts
// Interface automatically generated by schemas-to-ts

export interface CallApiHistory {
  id: number;
  attributes: {
    createdAt: Date;    updatedAt: Date;    publishedAt?: Date;    title?: string;
    method?: string;
    url?: string;
    request_body?: any;
    response_body?: any;
    status?: number;
  };
}
export interface CallApiHistory_Plain {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  title?: string;
  method?: string;
  url?: string;
  request_body?: any;
  response_body?: any;
  status?: number;
}

export interface CallApiHistory_NoRelations {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  title?: string;
  method?: string;
  url?: string;
  request_body?: any;
  response_body?: any;
  status?: number;
}

export interface CallApiHistory_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  title?: string;
  method?: string;
  url?: string;
  request_body?: any;
  response_body?: any;
  status?: number;
}